FIX: warnings

+++ Examples/Segmentation/HoughTransform2DCirclesImageFilter.cxx
@@ -218,8 +218,8 @@ int main( int argc, char *argv[] )
   // Software Guide : BeginCodeSnippet
     for(double angle = 0;angle <= 2*PI;angle+=PI/60)
       {
-      m_Index[0]=(*it_circles)->GetObjectToParentTransform()->GetOffset()[0] + (*it_circles)->GetRadius()[0]*cos(angle);
-      m_Index[1]=(*it_circles)->GetObjectToParentTransform()->GetOffset()[1] + (*it_circles)->GetRadius()[0]*sin(angle);
+      m_Index[0]=(long int)((*it_circles)->GetObjectToParentTransform()->GetOffset()[0] + (*it_circles)->GetRadius()[0]*cos(angle));
+      m_Index[1]=(long int)((*it_circles)->GetObjectToParentTransform()->GetOffset()[1] + (*it_circles)->GetRadius()[0]*sin(angle));
 
       if(m_Index[0]>=0 && m_Index[0]<(long)m_OutputImage->GetLargestPossibleRegion().GetSize()[0]
          && m_Index[1]>=0 && m_Index[1]<(long)m_OutputImage->GetLargestPossibleRegion().GetSize()[1])
